# docs/swagger.yaml
openapi: 3.0.3

info:
  title: CloudLite API
  description: REST API for the CloudLite file-storage backend.
  version: 0.1.0

servers:
  - url: http://localhost:8000
    description: Local development server

security:
  - bearerAuth: []

tags:
  - name: Folders
    description: Virtual folder hierarchy endpoints currently implemented by the backend.
  - name: Upload Sessions
    description: Batch-upload session lifecycle (create -> stage files -> commit or cancel)
  - name: Change Logs
    description: User-scoped change log sync endpoint.

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------

paths:

  # ---------------------------------------------------------------------------
  # Folders
  # ---------------------------------------------------------------------------

  /api/folders:
    get:
      tags: [Folders]
      summary: List folders
      description: >
        Returns the authenticated user's root folders when `parentId` is omitted,
        or the direct child folders of the provided parent folder. The parent folder,
        when specified, must belong to the requesting user.
      operationId: listFolders
      parameters:
        - $ref: '#/components/parameters/ParentFolderId'
      responses:
        '200':
          description: Folder list returned successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Folders]
      summary: Create a folder
      description: >
        Creates a new folder for the authenticated user.
        The folder name must be unique within the given parent (or among root folders
        if `parentId` is omitted). The parent folder, when specified, must exist and
        belong to the requesting user.
      operationId: createFolder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderRequest'
      responses:
        '201':
          description: Folder created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FolderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/folders/batch:
    post:
      tags: [Folders]
      summary: Create folders as a batch tree
      description: >
        Creates multiple folders for the authenticated user from a nested directory
        tree. The backend links each child folder to its parent and computes each
        full folder path before persisting the batch. Folder names must be unique
        among siblings and cannot contain path separators.
      operationId: createFoldersBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFolderBatchRequest'
      responses:
        '201':
          description: Folders created successfully in pre-order traversal.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FolderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # Change Logs
  # ---------------------------------------------------------------------------

  /api/change-logs:
    get:
      tags: [Change Logs]
      summary: Get change logs since a synced ID
      description: >
        Returns all change log entries for the authenticated user with an ID greater
        than `latestSyncedId`, ordered by ID ascending. Clients can pass the latest
        log ID they have already synchronized and receive only newer changes.
      operationId: getChangeLogsSince
      parameters:
        - $ref: '#/components/parameters/LatestSyncedId'
      responses:
        '200':
          description: User-scoped change log entries newer than the provided ID.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChangeLogResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # Upload Sessions
  # ---------------------------------------------------------------------------

  /api/upload/sessions:
    post:
      tags: [Upload Sessions]
      summary: Create an upload session
      description: >
        Opens a new upload session targeting the specified folder.
        All files staged in this session will be published into that folder on commit.
      operationId: createSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
      responses:
        '201':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/upload/sessions/{sessionId}/files/batch:
    post:
      tags: [Upload Sessions]
      summary: Stage one or more files
      description: >
        Streams one or more files into the session in a single `multipart/form-data`
        request. All duplicate-name checks are performed with a single database query
        before any storage writes, so the request is rejected atomically if any
        filename is already staged in the session or appears more than once in the
        request. Files are then streamed individually to the storage backend without
        being buffered in heap memory.
        If a storage or persistence error occurs after some files have been written,
        the already-stored blobs are cleaned up on a best-effort basis.
      operationId: stageFiles
      parameters:
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: One or more binary file parts.
      responses:
        '201':
          description: All files staged successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StagedFileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/upload/sessions/{sessionId}/commit:
    post:
      tags: [Upload Sessions]
      summary: Commit an upload session
      description: >
        Atomically publishes all staged files into the target folder.
        The session status changes to `COMMITTED`. Each staged filename must be
        unique within the target folder; if any conflict is detected the entire
        commit is rejected and the session remains `OPEN`.
      operationId: commitSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session committed; returns the newly created file records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublishedFileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /api/upload/sessions/{sessionId}:
    delete:
      tags: [Upload Sessions]
      summary: Cancel an upload session
      description: >
        Cancels an open session and discards all staged metadata.
        The session status changes to `CANCELLED`; staged blob files are left for
        later cleanup.
      operationId: cancelSession
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session cancelled successfully. No content returned.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

# ---------------------------------------------------------------------------
# Components
# ---------------------------------------------------------------------------

components:

  # -------------------------------------------------------------------------
  # Security schemes
  # -------------------------------------------------------------------------

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak-issued JWT. Obtain a token from the Keycloak realm
        (`/realms/development/protocol/openid-connect/token`) and supply it
        as `Authorization: Bearer <token>`.

  # -------------------------------------------------------------------------
  # Reusable path parameters
  # -------------------------------------------------------------------------

  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      description: UUID of the upload session.
      schema:
        type: string
        format: uuid
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
    FolderId:
      name: folderId
      in: path
      required: true
      description: UUID of the folder.
      schema:
        type: string
        format: uuid
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
    ParentFolderId:
      name: parentId
      in: query
      required: false
      description: Parent folder UUID. Omit to list root folders.
      schema:
        type: string
        format: uuid
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
    LatestSyncedId:
      name: latestSyncedId
      in: query
      required: false
      description: Latest change log ID already synchronized by the client. Defaults to 0.
      schema:
        type: integer
        format: int64
        minimum: 0
        default: 0
        example: 42

  # -------------------------------------------------------------------------
  # Schemas (DTOs)
  # -------------------------------------------------------------------------

  schemas:

    # -- Request bodies -------------------------------------------------------

    CreateFolderRequest:
      type: object
      required: [name]
      description: Request body for creating a new folder.
      properties:
        name:
          type: string
          minLength: 1
          description: Folder name. Must be unique within the parent (or among root folders).
          example: Documents
        parentId:
          type: string
          format: uuid
          nullable: true
          description: UUID of the parent folder. Omit to create a root folder.
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7

    CreateFolderBatchRequest:
      type: object
      required: [directories]
      description: Request body for creating a batch of folders as a nested tree.
      properties:
        parentId:
          type: string
          format: uuid
          nullable: true
          description: Existing parent folder ID. Omit to create root folders.
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        directories:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateFolderTreeNodeRequest'

    CreateFolderTreeNodeRequest:
      type: object
      required: [name]
      description: Folder node in a batch creation tree.
      properties:
        name:
          type: string
          minLength: 1
          description: Folder name. Must be unique among sibling nodes.
          example: Photos
        children:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/CreateFolderTreeNodeRequest'

    CreateSessionRequest:
      type: object
      required: [targetFolderId]
      description: Request body for creating a new upload session.
      properties:
        targetFolderId:
          type: string
          format: uuid
          description: UUID of the destination folder where files will be published on commit.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6

    # -- Response objects -----------------------------------------------------

    FolderResponse:
      type: object
      required: [id, name, path, ownerId, createdAt, updatedAt]
      description: Metadata for a virtual folder.
      properties:
        id:
          type: string
          format: uuid
          description: Unique folder identifier.
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        name:
          type: string
          description: Folder name.
          example: Documents
        path:
          type: string
          description: Full folder path.
          example: /Documents/Photos
        parentId:
          type: string
          format: uuid
          nullable: true
          description: UUID of the parent folder, or null for root folders.
          example: null
        ownerId:
          type: string
          format: uuid
          description: UUID of the folder owner.
          example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp when the folder was created.
          example: '2026-04-25T20:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp of the last update.
          example: '2026-04-25T20:00:00Z'

    SessionResponse:
      type: object
      required: [id, targetFolderId, status, createdAt]
      description: Metadata of an upload session.
      properties:
        id:
          type: string
          format: uuid
          description: Unique session identifier.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        targetFolderId:
          type: string
          format: uuid
          description: UUID of the destination folder.
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        status:
          $ref: '#/components/schemas/UploadSessionStatus'
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp when the session was created.
          example: '2026-04-25T20:00:00Z'

    StagedFileResponse:
      type: object
      required: [id, fileName, sizeBytes, sha256, status, createdAt]
      description: Metadata for a file that has been staged inside an upload session.
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the staged-file record.
          example: 1b9d6bcd-bbfd-4b2d-9b5d-ab8dfbbd4bed
        fileName:
          type: string
          description: Name the file will have after the session is committed.
          example: report.pdf
        sizeBytes:
          type: integer
          format: int64
          description: File size in bytes.
          example: 204800
        sha256:
          type: string
          description: Hex-encoded SHA-256 digest of the file content.
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        status:
          $ref: '#/components/schemas/UploadSessionFileStatus'
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp when the file was staged.
          example: '2026-04-25T20:01:00Z'

    PublishedFileResponse:
      type: object
      required: [id, name, folderId, blobId, sizeBytes, sha256, createdAt]
      description: Metadata for a file that has been published into a folder after session commit.
      properties:
        id:
          type: string
          format: uuid
          description: Unique file identifier.
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          description: Filename inside the folder.
          example: report.pdf
        folderId:
          type: string
          format: uuid
          description: UUID of the parent folder.
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        blobId:
          type: string
          format: uuid
          description: UUID of the underlying blob record.
          example: 9b2c3d4e-5f6a-7b8c-9d0e-1f2a3b4c5d6e
        sizeBytes:
          type: integer
          format: int64
          description: File size in bytes.
          example: 204800
        sha256:
          type: string
          description: Hex-encoded SHA-256 digest of the file content.
          example: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
        mimeType:
          type: string
          nullable: true
          description: MIME type of the file, or null if unknown.
          example: application/pdf
        createdAt:
          type: string
          format: date-time
          description: ISO-8601 timestamp when the file was created.
          example: '2026-04-25T20:02:00Z'

    ChangeLogResponse:
      type: object
      required: [id, eventType, entityType, timestamp, userId]
      description: User-scoped change log entry for client synchronization.
      properties:
        id:
          type: integer
          format: int64
          description: Monotonic change log identifier.
          example: 43
        eventType:
          $ref: '#/components/schemas/EventType'
        entityType:
          $ref: '#/components/schemas/EntityType'
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 timestamp when the change was logged.
          example: '2026-04-25T20:03:00Z'
        fileId:
          type: string
          format: uuid
          nullable: true
          description: Changed file ID, or null for folder-only events.
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        folderId:
          type: string
          format: uuid
          nullable: true
          description: Changed folder ID, or null for file-only events.
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        userId:
          type: string
          format: uuid
          description: User that performed the change.
          example: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11

    # -- Enums ----------------------------------------------------------------

    UploadSessionStatus:
      type: string
      description: Lifecycle status of an upload session.
      enum: [OPEN, COMMITTED, CANCELLED, FAILED]
      example: OPEN

    UploadSessionFileStatus:
      type: string
      description: Lifecycle status of a file within an upload session.
      enum: [UPLOADED, COMMITTED, FAILED]
      example: UPLOADED

    EventType:
      type: string
      description: Type of change that was logged.
      enum: [CREATE, RENAME, MOVE, DELETE]
      example: CREATE

    EntityType:
      type: string
      description: Type of entity that changed.
      enum: [FILE, DIRECTORY]
      example: FILE

    # -- Error ----------------------------------------------------------------

    ErrorResponse:
      type: object
      required: [status, error, message]
      description: Standard error body returned by GlobalExceptionHandler for 4xx/5xx responses.
      properties:
        status:
          type: integer
          description: HTTP status code.
          example: 404
        error:
          type: string
          description: HTTP reason phrase.
          example: Not Found
        message:
          type: string
          description: Human-readable description of what went wrong.
          example: "Folder not found: 3fa85f64-5717-4562-b3fc-2c963f66afa6"

  # -------------------------------------------------------------------------
  # Reusable responses
  # -------------------------------------------------------------------------

  responses:
    BadRequest:
      description: The request was malformed or violated a business rule.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: No valid Bearer token was provided.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated user does not own the requested resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: A resource with the same name already exists in the target location.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'